# api/openapi.yaml
openapi: 3.1.0
info:
  title: Server Crawl Stars REST API
  version: 0.1.0
  description: 클라이언트 연동 테스트용 REST API입니다. 매칭 join, room 상태 확인, debug room 생성/시작/삭제를 제공합니다.
...
servers:
  - url: /
    description: 현재 접속한 서버 origin
  - url: http://localhost:8080
    description: 로컬 개발 서버
...
paths:
...
  /matchmaking/join:
    post:
      tags: [Matchmaking]
      summary: 매칭에 참가하고 WebSocket 경로 받기
      description: |
        요청한 gameMode와 같은 여유 waiting room을 찾거나 새 room을 만들고 player를 발급합니다.
        body가 없거나, 빈 object이거나, gameMode가 빈 문자열이면 기본값 duel_1v1을 선택합니다.
        새 client는 optional `characterType`에 stable ID 0=Shelly, 1=Colt, 2=Lily 중 하나를 명시합니다. SL-82에서는 legacy client의 field 생략만 Shelly 0으로 보정하고 structured warning을 한 번 기록합니다. explicit null, non-integer, string/bool/object/array, 지원하지 않는 정수는 400 `invalid_character_type`입니다. SL-98에서 field를 required로 전환합니다.
        Raw request body가 1024 bytes를 초과하면 400 `invalid_request`로 거부합니다.

        첫 번째 참가자는 `status: waiting` 상태를 받습니다. 이 상태에서는 WebSocket input은 받을 수 있지만 gameplay snapshot은 broadcast하지 않습니다.
        첫 human matchmaking join부터 10초인 room-owned one-shot deadline을 시작합니다. 후속 human join과 partial manual bot 추가는 deadline을 reset하지 않습니다.
        deadline을 먼저 획득하면 selected mode의 남은 participant slot을 bot으로 충원하고 room을 matched로 잠급니다. 같은 순간 late join은 다른 waiting room을 찾거나 만들며, active-room cap이면 room_cap_reached 409를 반환합니다.
        선택 mode의 participant capacity가 차도 REST 응답 shape는 유지하며 room은 `status: waiting`으로 남습니다. Room participant에는 human과 server-owned bot이 함께 들어갈 수 있습니다.
        연결된 human WebSocket session이 attach quorum을 만족하면 각 human이 `Type: Ready` 이벤트로 bot을 포함한 full participant의 spawn 정보를 받습니다. human session만 Ready ACK를 보내면 server가 countdown 시작 신호를 1번 보냅니다. Bot은 token, WebSocket, Ready ACK가 없습니다. Client는 fake timer를 표시하고, server는 5초를 내부에서 센 뒤 room을 `started`로 바꾸고 room config의 tick rate로 snapshot broadcast를 시작합니다.
        1명만으로 디버그하려면 `POST /rooms/{roomID}/start`를 호출합니다.
      operationId: joinMatchmaking
      requestBody:
        description: |
          선택할 game mode입니다. Request body 자체는 optional입니다.
          body 없음, 빈 object, gameMode 빈 문자열은 모두 duel_1v1로 처리합니다.
          새 client는 canonical game mode ID와 optional `characterType` stable ID 0/1/2를 보냅니다. characterType 생략은 SL-82 legacy compatibility로만 Shelly 0이 되며 SL-98에서 required가 됩니다.
          Raw request body의 최대 크기는 1024 bytes이며 초과하면 `invalid_request`입니다.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchmakingJoinRequest"
            examples:
              default:
                summary: 기본 duel mode
                value:
                  characterType: 0
              solo:
                summary: 6인 solo mode
                value:
                  gameMode: solo
                  characterType: 1
              team:
                summary: 3 대 3 team mode
                value:
                  gameMode: team
                  characterType: 2
              legacyMissingCharacterType:
                summary: legacy field 생략은 SL-82에서만 Shelly로 보정
                value:
                  gameMode: duel_1v1
      responses:
        "201":
          description: 선택된 gameMode, 발급된 room, player, player session token, WebSocket 경로
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchmakingJoin"
              example:
                gameMode: duel_1v1
                room:
                  id: room_AbCdEfGhIjKlMnOpQrStUv
                  gameMode: duel_1v1
                  status: waiting
                  players:
                    - id: player_VuTsRqPoNmLkJiHgFeDcBa
                      team: red
                      slot: 0
                      isBot: false
                      characterType: 1
                  maxPlayers: 6
                  map:
                    width: 5
                    height: 5
                    index: 0
                    maxPlayers: 6
                    tileSize: 1.2
                    map: [[1, 1, 1, 1, 1], [1, 0, 0, 0, 1], [1, 0, 1, 0, 1], [1, 0, 0, 0, 1], [1, 1, 1, 1, 1]]
                  latestSnapshot:
                    tick: 0
                    playerCount: 1
                    projectileCount: 0
                player:
                  id: player_VuTsRqPoNmLkJiHgFeDcBa
                  team: red
                  slot: 0
                  isBot: false
                  characterType: 1
                sessionToken: example_redacted_session_token_000000000000
                webSocketPath: "/rooms/room_AbCdEfGhIjKlMnOpQrStUv/players/player_VuTsRqPoNmLkJiHgFeDcBa?token=example_redacted_session_token_000000000000"
        "400":
          description: 요청 JSON이 잘못됐거나 raw body가 1024 bytes를 초과했거나 지원하지 않는 gameMode/characterType입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidGameMode:
                  summary: 지원하지 않는 game mode
                  value:
                    error:
                      code: invalid_game_mode
                      message: invalid game mode
                invalidRequest:
                  summary: malformed JSON request
                  value:
                    error:
                      code: invalid_request
                      message: invalid request
                invalidCharacterType:
                  summary: 지원하지 않거나 잘못된 character type
                  value:
                    error:
                      code: invalid_character_type
                      message: invalid character type
        "409":
          description: Timer-first late join을 포함해 active-room cap에 도달하면 `room_cap_reached`를 반환합니다. 테스트 중이면 `DELETE /rooms`로 debug room을 비울 수 있습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: |
            Process-local per-IP token bucket이 요청을 거부했습니다. 응답 code는 `rate_limited`입니다.
            Quota는 request body decode와 store join보다 먼저 평가하므로 429가 request 400, room cap 409, 발급 실패 500보다 우선합니다. Store에서 409/500으로 끝난 허용 요청도 quota를 이미 소비합니다.
          headers:
            Retry-After:
              description: 다시 요청하기 전 대기 시간을 올림한 최소 1초의 delta-seconds 정수입니다.
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: rate_limited
                  message: rate limit exceeded
        "500":
          description: 안전한 opaque ID 또는 session token 발급에 실패했습니다. 응답 code는 `internal_error`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms:
    get:
      tags: [Rooms]
      summary: active debug room 목록 보기
      description: |
        room 상태, player 목록, 서버가 사용하는 map 데이터, 최신 snapshot 요약을 확인합니다.
        Debug API는 기본 비활성화이며, 비활성화 상태에서는 route 자체가 없는 것처럼 404를 반환합니다.
      operationId: listRooms
      security:
        - DebugBearer: []
      responses:
        "200":
          description: active room 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomList"
        "401":
          description: Debug API가 활성화됐지만 Bearer credential이 없거나 올바르지 않습니다. 응답 code는 `unauthorized`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Debug API는 기본 비활성화이며, 이때 JSON code `not_found`로 응답합니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Rooms]
      summary: debug room 직접 만들기
      description: |
        매칭을 거치지 않고 수동 검증용 room을 만듭니다.
        Debug API는 기본 비활성화이며, 비활성화 상태에서는 route 자체가 없는 것처럼 404를 반환합니다.
      operationId: createRoom
      security:
        - DebugBearer: []
      responses:
        "201":
          description: 생성된 room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          description: Debug API가 활성화됐지만 Bearer credential이 없거나 올바르지 않습니다. 응답 code는 `unauthorized`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Debug API는 기본 비활성화이며, 이때 JSON code `not_found`로 응답합니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Active room cap에 도달했습니다. `DELETE /rooms`로 debug room을 비울 수 있습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 안전한 opaque room ID 발급에 실패했습니다. 응답 code는 `internal_error`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Rooms]
      summary: 모든 debug room 비우기
      description: |
        현재 메모리에 있는 모든 room을 삭제하고, 실행 중인 ticker와 WebSocket 연결을 닫습니다.
        `room_cap_reached` 409가 난 뒤 테스트 환경을 즉시 회복할 때 사용합니다.
        Debug API는 기본 비활성화이며, 비활성화 상태에서는 route 자체가 없는 것처럼 404를 반환합니다.
      operationId: clearRooms
      security:
        - DebugBearer: []
      responses:
        "200":
          description: 삭제된 room 수
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearRoomsResult"
        "401":
          description: Debug API가 활성화됐지만 Bearer credential이 없거나 올바르지 않습니다. 응답 code는 `unauthorized`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Debug API는 기본 비활성화이며, 이때 JSON code `not_found`로 응답합니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{roomID}:
    get:
      tags: [Rooms]
      summary: debug room 상세 보기
      description: |
        room 상태, player 목록, 서버 map 데이터, 최신 snapshot 요약을 확인합니다.
        Debug API는 기본 비활성화이며, 비활성화 상태에서는 route 자체가 없는 것처럼 404를 반환합니다.
      operationId: getRoom
      security:
        - DebugBearer: []
      parameters:
        - $ref: "#/components/parameters/RoomID"
      responses:
        "200":
          description: room 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          description: Debug API가 활성화됐지만 Bearer credential이 없거나 올바르지 않습니다. 응답 code는 `unauthorized`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Debug API가 기본 비활성화면 JSON code `not_found`, 활성화되고 인증됐지만 room이 없으면 `room_not_found`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Rooms]
      summary: debug room 하나 삭제
      description: |
        지정한 room을 삭제하고, 해당 room의 ticker와 WebSocket 연결을 닫습니다.
        Debug API는 기본 비활성화이며, 비활성화 상태에서는 route 자체가 없는 것처럼 404를 반환합니다.
      operationId: deleteRoom
      security:
        - DebugBearer: []
      parameters:
        - $ref: "#/components/parameters/RoomID"
      responses:
        "200":
          description: 삭제된 room 수
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearRoomsResult"
        "401":
          description: Debug API가 활성화됐지만 Bearer credential이 없거나 올바르지 않습니다. 응답 code는 `unauthorized`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Debug API가 기본 비활성화면 JSON code `not_found`, 활성화되고 인증됐지만 room이 없으면 `room_not_found`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{roomID}/players:
    post:
      tags: [Rooms]
      summary: debug room에 player 발급
      description: |
        수동 room 검증용 player를 추가하고, 최초 응답에서만 raw player session token을 발급합니다. 현재 fixture 기준 최대 player 수는 6명입니다.
        Matchmaking room은 selected mode의 participant capacity가 human과 bot으로 모두 차 lifecycle이 잠기면 추가 player를 받지 않습니다.
        Debug API는 기본 비활성화이며, 비활성화 상태에서는 route 자체가 없는 것처럼 404를 반환합니다.
      operationId: createRoomPlayer
      security:
        - DebugBearer: []
      parameters:
        - $ref: "#/components/parameters/RoomID"
      responses:
        "201":
          description: 발급된 player, player session token, WebSocket 경로
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSessionResponse"
        "401":
          description: Debug API가 활성화됐지만 Bearer credential이 없거나 올바르지 않습니다. 응답 code는 `unauthorized`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Debug API가 기본 비활성화면 JSON code `not_found`, 활성화되고 인증됐지만 room이 없으면 `room_not_found`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: room player 수가 map의 maxPlayers에 도달했거나 matchmaking lifecycle이 이미 잠겼습니다. 응답 code는 `room_full`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 안전한 opaque player ID 또는 session token 발급에 실패했습니다. 응답 code는 `internal_error`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rooms/{roomID}/start:
    post:
      tags: [Rooms]
      summary: debug room 시뮬레이션 시작
      description: |
        player가 1명 이상 있는 room을 `started`로 바꾸고 30Hz snapshot broadcast를 시작합니다.
        매칭 flow에서는 선택 mode의 participant capacity가 채워진 뒤 연결된 모든 human WebSocket client가 Ready ACK를 보내고 server 내부 5초 countdown이 끝나면 자동으로 같은 전이가 일어납니다.
        Debug API는 기본 비활성화이며, 비활성화 상태에서는 route 자체가 없는 것처럼 404를 반환합니다.
      operationId: startRoom
      security:
        - DebugBearer: []
      parameters:
        - $ref: "#/components/parameters/RoomID"
      responses:
        "200":
          description: 시작된 room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          description: Debug API가 활성화됐지만 Bearer credential이 없거나 올바르지 않습니다. 응답 code는 `unauthorized`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Debug API가 기본 비활성화면 JSON code `not_found`, 활성화되고 인증됐지만 room이 없으면 `room_not_found`입니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: player가 없어 시작할 수 없습니다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    DebugBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque debug token
      description: |
        `ENABLE_DEBUG_API=true`일 때만 사용하는 운영용 debug credential입니다.
        `Authorization: Bearer <DEBUG_API_TOKEN>`으로 전달하며 Swagger UI는 값을 저장하지 않습니다.
        Authorization header가 없거나, 값이 틀리거나, header가 여러 개면 route dispatch 전에 401 `unauthorized`입니다. Debug API가 비활성화됐으면 credential과 관계없이 404 `not_found`입니다.
  parameters:
    RoomID:
      name: roomID
      in: path
      required: true
      description: "`POST /rooms` 또는 `POST /matchmaking/join`으로 받은 room ID"
      schema:
        $ref: "#/components/schemas/OpaqueRoomID"
  schemas:
    OpaqueRoomID:
      type: string
      description: 16 random bytes를 Raw URL Base64로 인코딩하고 `room_` prefix를 붙인 opaque ID입니다.
      pattern: "^room_[A-Za-z0-9_-]{22}$"
      example: room_AbCdEfGhIjKlMnOpQrStUv
...
    PlayerSessionToken:
      type: string
      description: |
        32 random bytes를 Raw URL Base64로 인코딩한 43자 token입니다.
        Raw 값은 player 발급 응답의 `sessionToken`과 tokenized `webSocketPath` 두 곳에 같은 secret으로 나타나고, 이후 WebSocket `token` query로 다시 전달됩니다.
        Token은 일회용이 아니며 room/player session이 존재하는 동안 재사용할 수 있습니다. Unmatched disconnect는 room-owned 10초 fill deadline과 credential을 유지합니다. matched/loading/starting disconnect는 pre-start cancel로 room을 삭제하고, started room은 TTL/hard lifetime 뒤 삭제됩니다. Failed upgrade는 room을 취소하지 않아 재시도할 수 있습니다.
        서버 private state에는 SHA-256 digest만 저장합니다. 세 secret-bearing surface 모두 log나 telemetry에 기록하면 안 됩니다.
      pattern: "^[A-Za-z0-9_-]{43}$"
...
    RoomList:
      type: object
      required: [rooms]
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
    MatchmakingJoinRequest:
      type: object
      description: |
        Matchmaking mode 선택 요청입니다. Canonical non-empty 값은 duel_1v1, solo, team입니다.
        Field 생략과 빈 문자열은 기존 client 호환을 위해 server에서 기본 duel_1v1로 처리합니다.
        characterType은 SL-82에서 optional migration field입니다. 새 client는 0=Shelly, 1=Colt, 2=Lily를 명시하고, legacy 생략만 Shelly 0으로 보정됩니다. SL-98에서 required로 전환합니다.
      properties:
        gameMode:
          type: string
          oneOf:
            - enum: [duel_1v1, solo, team]
            - const: ""
              title: Legacy empty value
          default: duel_1v1
        characterType:
          $ref: "#/components/schemas/CharacterType"
    CharacterType:
      type: integer
      description: 캐릭터의 stable numeric ID입니다. 0=Shelly, 1=Colt, 2=Lily이며 재번호화하지 않습니다.
      enum: [0, 1, 2]
      example: 1
    MatchmakingJoin:
      type: object
      required: [gameMode, room, player, sessionToken, webSocketPath]
      properties:
        gameMode:
          type: string
          description: 선택되어 room에 고정된 game mode ID입니다. nested room.gameMode와 항상 같습니다.
          enum: [duel_1v1, solo, team]
        room:
          $ref: "#/components/schemas/Room"
        player:
          $ref: "#/components/schemas/HumanPlayer"
        sessionToken:
          $ref: "#/components/schemas/PlayerSessionToken"
        webSocketPath:
          type: string
          description: 발급된 raw session token을 `token` query parameter로 포함한 연결 경로입니다. 예시는 의도적으로 redacted했습니다.
          example: "/rooms/room_AbCdEfGhIjKlMnOpQrStUv/players/player_VuTsRqPoNmLkJiHgFeDcBa?token=<player-session-token>"
    PlayerSessionResponse:
      type: object
      required: [player, sessionToken, webSocketPath]
      description: Debug player 발급 응답입니다. Server가 raw session token을 내보내는 발급 응답이며, 같은 secret이 `sessionToken`과 tokenized `webSocketPath` 두 field에 나타납니다. 공개 Room/Player payload에는 포함되지 않습니다.
      properties:
        player:
          $ref: "#/components/schemas/HumanPlayer"
        sessionToken:
          $ref: "#/components/schemas/PlayerSessionToken"
        webSocketPath:
          type: string
          description: 발급된 raw session token을 `token` query parameter로 포함한 연결 경로입니다. 예시는 의도적으로 redacted했습니다.
          example: "/rooms/room_AbCdEfGhIjKlMnOpQrStUv/players/player_VuTsRqPoNmLkJiHgFeDcBa?token=<player-session-token>"
    Room:
      type: object
      required: [id, gameMode, status, players, maxPlayers, map, latestSnapshot]
      description: room 상태 확인 응답입니다. `gameMode`는 room 생성 시 선택되어 lifecycle 동안 바뀌지 않습니다. `map`은 클라이언트가 같은 tile grid를 그릴 때 사용하며 row는 Base64 문자열이 아니라 JSON number array로 직렬화됩니다. 기본 source는 server binary가 embed한 `server-config/game-config.json`의 map이고, config 로드나 검증에 실패하면 `StaticGameConfig()`의 5x5 map으로 fallback합니다.
      properties:
        id:
          $ref: "#/components/schemas/OpaqueRoomID"
        gameMode:
          type: string
          description: 이 room이 고정해서 소유하는 selected game mode ID입니다.
          enum: [duel_1v1, solo, team]
        status:
          type: string
          description: >-
            waiting이면 REST room lifecycle상 아직 gameplay snapshot broadcast 전,
            started이면 30Hz snapshot broadcast 중입니다. Match render 준비는
            WebSocket `Type: Ready`, countdown 시작은 1회성
            `Snapshot.status: starting`에서 봅니다.
          enum: [waiting, started]
        players:
          type: array
          description: human과 server-owned bot을 포함할 수 있는 room participant 목록입니다.
          items:
            $ref: "#/components/schemas/Player"
        maxPlayers:
          type: integer
          description: 현재 room map fixture가 허용하는 최대 player 수입니다.
          minimum: 1
          example: 6
        map:
          $ref: "#/components/schemas/MapData"
        latestSnapshot:
          $ref: "#/components/schemas/SnapshotSummary"
    Player:
      type: object
      description: Room에 참여하는 generic participant입니다. Human과 server-owned bot을 모두 나타낼 수 있습니다.
      required: [id, team, slot, isBot, characterType]
      properties:
        id:
          $ref: "#/components/schemas/OpaquePlayerID"
        team:
          type: string
          description: 서버 assignment helper가 room-local mode/team rule 기준으로 발급한 team입니다. duel/team은 red/blue, solo는 solo-1부터 solo-6을 사용합니다.
          enum: [red, blue, solo-1, solo-2, solo-3, solo-4, solo-5, solo-6]
        slot:
          type: integer
          description: team 안에서의 slot입니다.
          minimum: 0
        isBot:
          type: boolean
          description: 서버가 소유해 session과 WebSocket credential이 없는 bot participant인지 나타냅니다.
        characterType:
          $ref: "#/components/schemas/CharacterType"
    HumanPlayer:
      allOf:
        - $ref: "#/components/schemas/Player"
        - type: object
          required: [isBot]
          properties:
            isBot:
              type: boolean
              const: false
              description: Credential 발급 응답은 human player만 반환합니다.
    SnapshotSummary:
      type: object
      required: [tick, playerCount, projectileCount]
      description: 마지막으로 생성된 snapshot의 짧은 요약입니다. 아직 tick 전이면 tick은 0입니다.
      properties:
        tick:
          type: integer
          minimum: 0
        playerCount:
          type: integer
          minimum: 0
        projectileCount:
          type: integer
          minimum: 0
...
    ClearRoomsResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: integer
          minimum: 0
          description: 삭제된 room 수입니다.
          example: 5
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - room_not_found
            - room_cap_reached
            - room_full
            - room_has_no_players
            - method_not_allowed
            - not_found
            - invalid_game_mode
            - invalid_request
            - invalid_character_type
            - unauthorized
            - rate_limited
            - internal_error
        message:
          type: string